requirements.txt
flask~=3.0.0
python-dotenv~=1.0.1
psycopg2~=2.9.9
openai~=1.16.2
gigachat~=0.1.14